Windows version now compiles! Also, several compilation warnings removed.
git-svn-id: https://klub.com.pl/var/svn/omnet@214 056025b2-ba0e-0410-831a-f040b6dfa49e

+++ wimax/wmaxmac.cc
@@ -82,7 +82,7 @@ void WMaxMac::stringUpdate() {
 	// count all messages in the queue
 
         stringstream displayIt;
-        displayIt << queuedMsgsCnt << "msgs in " << Conns.size() << " queues.";
+        displayIt << queuedMsgsCnt << "msgs in " << (int)Conns.size() << " queues.";
         displayString().setTagArg("t",0, (displayIt.str()).c_str());
     }
 }
@@ -1149,7 +1149,7 @@ void WMaxMacSS::schedule(WMaxMsgUlMap * ulmap)
                 }
 
                 BEpoint++;
-                BEpoint = BEpoint%Conns.size();
+                BEpoint = BEpoint%(int)Conns.size();
 
                 if (BEit->type == WMAX_CONN_TYPE_BE) {
                     if (BEit->qos.be.reqbw) {
